rename and refactor sjf schedulers
